+++ pkg/providers/gtm/provider.go
@@ -7,51 +7,82 @@ import (
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v7/pkg/gtm"
 	"github.com/akamai/terraform-provider-akamai/v5/pkg/meta"
 	"github.com/akamai/terraform-provider-akamai/v5/pkg/subprovider"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 type (
-	// Subprovider gathers gtm resources and data sources
-	Subprovider struct {
+	// PluginSubprovider gathers gtm resources and data sources
+	PluginSubprovider struct {
 		client gtm.GTM
 	}
 
-	option func(p *Subprovider)
+	// FrameworkSubprovider gathers property resources and data sources written using terraform-plugin-framework
+	FrameworkSubprovider struct {
+		client gtm.GTM
+	}
 )
+
+var _ subprovider.Framework = &FrameworkSubprovider{}
+var _ subprovider.Plugin = &PluginSubprovider{}
 
 var (
-	once sync.Once
+	oncePlugin, onceFramework sync.Once
 
-	inst *Subprovider
+	inst *PluginSubprovider
+
+	frameworkInst *FrameworkSubprovider
 )
 
-var _ subprovider.Plugin = &Subprovider{}
-
-// NewSubprovider returns a core sub provider
-func NewSubprovider() *Subprovider {
-	once.Do(func() {
-		inst = &Subprovider{}
+// NewSubprovider returns a new GTM subprovider
+func NewSubprovider() *PluginSubprovider {
+	oncePlugin.Do(func() {
+		inst = &PluginSubprovider{}
 	})
 
 	return inst
 }
 
-func withClient(c gtm.GTM) option {
-	return func(p *Subprovider) {
-		p.client = c
-	}
+// NewFrameworkSubprovider returns a core Framework based sub provider
+func NewFrameworkSubprovider() *FrameworkSubprovider {
+	onceFramework.Do(func() {
+		frameworkInst = &FrameworkSubprovider{}
+	})
+
+	return frameworkInst
 }
 
-// Client returns the DNS interface
-func (p *Subprovider) Client(meta meta.Meta) gtm.GTM {
+// Client returns the GTM interface
+func (p *PluginSubprovider) Client(meta meta.Meta) gtm.GTM {
 	if p.client != nil {
 		return p.client
 	}
 	return gtm.Client(meta.Session())
 }
 
+// Client returns the GTM interface
+func (f *FrameworkSubprovider) Client(meta meta.Meta) gtm.GTM {
+	if f.client != nil {
+		return f.client
+	}
+	return gtm.Client(meta.Session())
+}
+
+// Resources returns the GTM resources implemented using terraform-plugin-framework
+func (f *FrameworkSubprovider) Resources() []func() resource.Resource {
+	return []func() resource.Resource{}
+}
+
+// DataSources returns the GTM data sources implemented using terraform-plugin-framework
+func (f *FrameworkSubprovider) DataSources() []func() datasource.DataSource {
+	return []func() datasource.DataSource{
+		NewGTMResourceDataSource,
+	}
+}
+
 // Resources returns terraform resources for gtm
-func (p *Subprovider) Resources() map[string]*schema.Resource {
+func (p *PluginSubprovider) Resources() map[string]*schema.Resource {
 	return map[string]*schema.Resource{
 		"akamai_gtm_domain":     resourceGTMv1Domain(),
 		"akamai_gtm_property":   resourceGTMv1Property(),
@@ -64,7 +95,7 @@ func (p *Subprovider) Resources() map[string]*schema.Resource {
 }
 
 // DataSources returns terraform data sources for gtm
-func (p *Subprovider) DataSources() map[string]*schema.Resource {
+func (p *PluginSubprovider) DataSources() map[string]*schema.Resource {
 	return map[string]*schema.Resource{
 		"akamai_gtm_datacenter":         dataSourceGTMDatacenter(),
 		"akamai_gtm_datacenters":        dataSourceGTMDatacenters(),