+++ pkg/providers/gtm/testdata/TestDataGTMResource/missing_domain.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_gtm_resource" "my_gtm_resource" {
+  resource_name = "resource1"
+}